refactor: the layout of the in app step editor
